Started pulling timer code out of dispatch for #7
