+++ src/event/dispatch.cpp
@@ -19,50 +19,11 @@ namespace msa { namespace event {
 		#define MSA_MODULE_HOOK(retspec, name, ...)		name,
 		#include "event/hooks.hpp"
 		#undef MSA_MODULE_HOOK
-	};
-
-	typedef std::chrono::high_resolution_clock chrono_clock;
-	typedef chrono_clock::time_point chrono_time;
-	
-	class Timer
-	{
-		public:
-			int16_t id;
-			std::chrono::milliseconds period;
-			std::chrono::high_resolution_clock::time_point last_fired;
-			bool recurring;
-			IArgs *event_args;
-			Topic event_topic;
-			Timer(std::chrono::milliseconds period, Topic topic, const IArgs &args, bool recurring) :
-				id(0),
-				period(period),
-				last_fired(chrono_clock::now()),
-				recurring(recurring),
-				event_args(args.copy()),
-				event_topic(topic)
-			{}
-			Timer(const Timer &other) :
-				id(other.id),
-				period(other.period),
-				last_fired(other.last_fired),
-				recurring(other.recurring),
-				event_args(other.event_args->copy()),
-				event_topic(other.event_topic)
-			{}
-			~Timer()
-			{
-				delete event_args;
-			}
-			Timer &operator=(const Timer &other)
-			{
-				delete event_args;
-				event_args = other.event_args->copy();
-				id = other.id;
-				period = other.period;
-				last_fired = other.last_fired;
-				event_topic = other.event_topic;
-				return *this;
-			}
+		{
+			#define MSA_MODULE_HOOK(retspec, name, ...)			name,
+			#include "event/timer_hooks.hpp"
+			#undef MSA_MODULE_HOOK
+		}
 	};
 
 	typedef struct handler_context_type {
@@ -82,10 +43,7 @@ namespace msa { namespace event {
 		std::map<Topic, EventHandler> handlers;
 		std::stack<HandlerContext *> interrupted;
 		int sleep_time;
-		std::chrono::milliseconds tick_resolution;
-		chrono_time last_tick_time;
-		std::map<int16_t, Timer*> timers;
-		msa::thread::Mutex timers_mutex;
+		TimerContext timer;
 		std::vector<msa::cmd::Command *> commands;
 	};
 
@@ -105,9 +63,6 @@ namespace msa { namespace event {
 	static void edt_dispatch_event(msa::Handle hdl, const Event *e);
 	static void edt_fire_timers(msa::Handle hdl, chrono_time now);
 	static void dispose_handler_context(HandlerContext *ctx, bool wait);
-
-	static void cmd_timer(msa::Handle hdl, const msa::cmd::ParamList &params, HandlerSync *const sync);
-	static void cmd_deltimer(msa::Handle hdl, const msa::cmd::ParamList &params, HandlerSync *const sync);
 
 	extern int init(msa::Handle hdl, const msa::cfg::Section &config)
 	{
@@ -192,6 +147,11 @@ namespace msa { namespace event {
 		return &HOOKS;
 	}
 
+	extern TimerContext *get_timer_context(msa::Handle msa)
+	{
+		return msa->event->timer;
+	}
+
 	extern void subscribe(msa::Handle msa, Topic t, EventHandler handler)
 	{
 		msa->event->handlers[t] = handler;
@@ -209,76 +169,14 @@ namespace msa { namespace event {
 		push_event(msa, e);
 	}
 
-	extern int16_t schedule(msa::Handle msa, time_t timestamp, const Topic topic, const IArgs &args)
-	{
-		time_t ref_time = time(NULL);
-		if (ref_time >= timestamp)
-		{
-			return -1;
-		}
-		return delay(msa, std::chrono::seconds(timestamp), topic, args);
-	}
-
-	extern int16_t delay(msa::Handle msa, std::chrono::milliseconds delay, const Topic topic, const IArgs &args)
-	{
-		Timer *t = new Timer(delay, topic, args, false);
-		msa::thread::mutex_lock(&msa->event->timers_mutex);
-		t->id = msa->event->timers.size();
-		msa->event->timers[t->id] = t;
-		msa::thread::mutex_unlock(&msa->event->timers_mutex);
-		msa::log::debug(msa, "Scheduled a " + topic_str(t->event_topic) + " event to fire in " + std::to_string(delay.count()) + "ms (id = " + std::to_string(t->id) + ")");
-		return t->id;
-	}
-	
-	extern int16_t add_timer(msa::Handle msa, std::chrono::milliseconds period, const Topic topic, const IArgs &args)
-	{
-		Timer *t = new Timer(period, topic, args, true);
-		msa::thread::mutex_lock(&msa->event->timers_mutex);
-		t->id = msa->event->timers.size();
-		msa->event->timers[t->id] = t;
-		msa::thread::mutex_unlock(&msa->event->timers_mutex);
-		msa::log::debug(msa, "Scheduled a " + topic_str(t->event_topic) + " event to fire every " + std::to_string(period.count()) + "ms (id = " + std::to_string(t->id) + ")");
-		return t->id;
-	}
-
-	extern void remove_timer(msa::Handle msa, int16_t id)
-	{
-		EventDispatchContext *ctx = msa->event;
-		msa::thread::mutex_lock(&ctx->timers_mutex);
-		if (ctx->timers.find(id) == ctx->timers.end())
-		{
-			msa::thread::mutex_unlock(&ctx->timers_mutex);
-			throw std::logic_error("no timer with ID: " + std::to_string(id));
-		}
-		Timer *t = ctx->timers[id];
-		ctx->timers.erase(id);
-		msa::thread::mutex_unlock(&ctx->timers_mutex);
-		delete t;
-		msa::log::debug(msa, "Removed timer ID " + std::to_string(id));
-		return;
-	}
-
-	extern void get_timers(msa::Handle msa, std::vector<int16_t> &list)
-	{
-		EventDispatchContext *ctx = msa->event;
-		msa::thread::mutex_lock(&ctx->timers_mutex);
-		std::map<int16_t, Timer*>::const_iterator iter;
-		for (iter = ctx->timers.begin(); iter != ctx->timers.end(); iter++)
-		{
-			list.push_back(iter->first);
-		}
-		msa::thread::mutex_unlock(&ctx->timers_mutex);
-	}
-
 	static int create_event_dispatch_context(EventDispatchContext **event)
 	{
 		EventDispatchContext *edc = new EventDispatchContext;
 		msa::thread::mutex_init(&edc->queue_mutex, NULL);
-		msa::thread::mutex_init(&edc->timers_mutex, NULL);
+		msa::thread::mutex_init(&edc->timer.mutex, NULL);
+		edc->timer.last_tick_time = chrono_time::min();
 		edc->current_handler = NULL;
-		edc->last_tick_time = chrono_time::min();
-		edc->commands.push_back(new msa::cmd::Command("TIMER", "It schedules a command to execute in the future", "time-ms command", "r", cmd_timer));
-		edc->commands.push_back(new msa::cmd::Command("DELTIMER", "It deletes a timer", "timer-id", cmd_deltimer)),
+		edc->commands = get_timer_commands();
 		*event = edc;
 		return 0;
 	}
@@ -305,7 +203,7 @@ namespace msa { namespace event {
 		config.check_range("TICK_RESOLUTION", sleep_time, 1000, false);
 		int tick_res = config.get_or("TICK_RESOLUTION", 10);
 		hdl->event->sleep_time = sleep_time;
-		hdl->event->tick_resolution = std::chrono::milliseconds(tick_res);
+		hdl->event->timer.tick_resolution = std::chrono::milliseconds(tick_res);
 	}
 
 	static void *edt_start(void *args)
@@ -339,18 +237,18 @@ namespace msa { namespace event {
 			dispose_handler_context(intr_ctx, true);
 		}
 		msa::thread::mutex_destroy(&hdl->event->queue_mutex);
-		msa::thread::mutex_destroy(&hdl->event->timers_mutex);
+		msa::thread::mutex_destroy(&hdl->event->timer.mutex);
 		while (!hdl->event->queue.empty())
 		{
 			const Event *e = hdl->event->queue.top();
 			hdl->event->queue.pop();
 			delete e;
 		}
-		auto timer_iter = hdl->event->timers.begin();
-		while (timer_iter != hdl->event->timers.end())
+		auto timer_iter = hdl->event->timer.list.begin();
+		while (timer_iter != hdl->event->timer.list.end())
 		{
 			Timer *t = timer_iter->second;
-			timer_iter = hdl->event->timers.erase(timer_iter);
+			timer_iter = hdl->event->timer.list.erase(timer_iter);
 			delete t;
 		}
 	}
@@ -383,9 +281,9 @@ namespace msa { namespace event {
 
 		// check if we need to do timing tasks
 		chrono_time now = chrono_clock::now();
-		if (edc->last_tick_time + edc->tick_resolution <= now)
+		if (edc->timer.last_tick_time + edc->timer.tick_resolution <= now)
 		{
-			edc->last_tick_time = now;
+			edc->timer.last_tick_time = now;
 			edt_fire_timers(hdl, now);
 		}
 	}
@@ -393,32 +291,23 @@ namespace msa { namespace event {
 	static void edt_fire_timers(msa::Handle hdl, chrono_time now)
 	{
 		EventDispatchContext *ctx = hdl->event;
-		msa::thread::mutex_lock(&ctx->timers_mutex);
+		msa::thread::mutex_lock(&ctx->timer.mutex);
 		std::map<int16_t, Timer*>::iterator iter = ctx->timers.begin();
 		while (iter != ctx->timers.end())
 		{
 			Timer *t = iter->second;
-			if (t->last_fired + t->period <= now)
+			if (t->ready(now))
 			{
-				int16_t id = iter->first;
-				generate(hdl, t->event_topic, *t->event_args);
-				msa::log::debug(hdl, "Fired timer " + std::to_string(id));
-				if (t->recurring)
-				{
-					t->last_fired = now;
-					iter++;
-				}
-				else
+				t->fire(now);
+				if (!t->recurring())
 				{
 					iter = ctx->timers.erase(iter);
 					delete t;
-					msa::log::debug(hdl, "Completed and removed timer " + std::to_string(id));
+					msa::log::debug(hdl, "Completed and removed timer " + std::to_string(iter->first));
+					continue;
 				}
 			}
-			else
-			{
-				iter++;
-			}
+			iter++;
 		}
 		msa::thread::mutex_unlock(&ctx->timers_mutex);
 	}
@@ -564,83 +453,6 @@ namespace msa { namespace event {
 		msa::thread::mutex_lock(&msa->event->queue_mutex);
 		msa->event->queue.push(e);
 		msa::thread::mutex_unlock(&msa->event->queue_mutex);
-	}
-
-	static void cmd_timer(msa::Handle hdl, const msa::cmd::ParamList &params, HandlerSync *const UNUSED(sync))
-	{
-		bool recurring = params.has_option('r');
-		if (params.arg_count() < 2)
-		{
-			msa::agent::say(hdl, "You gotta give me a time and a command to execute, $USER_TITLE.");
-			return;
-		}
-		int period = 0;
-		try
-		{
-			period = std::stoi(params[0]);
-		}
-		catch (std::exception &e)
-		{
-			msa::agent::say(hdl, "Sorry, $USER_TITLE, but '" + params[0] + "' isn't a number of milliseconds.");
-			return;
-		}
-		if (period < 0)
-		{
-			msa::agent::say(hdl, "Sorry, $USER_TITLE, I might be good but I can't go back in time.");
-			msa::agent::say(hdl, "Please give me a positive number of milliseconds.");
-			return;
-		}
-		auto ms = std::chrono::milliseconds(period);
-		std::string cmd_str = "";
-		for (size_t i = 1; i < params.arg_count(); i++)
-		{
-			cmd_str += params[i];
-			if (i + 1 < params.arg_count())
-			{
-				cmd_str += " ";
-			}
-		}
-		std::string plural = ms.count() != 1 ? "s" : "";
-		std::string type = recurring ? "every" : "in";
-		int16_t id = -1;
-		if (recurring)
-		{
-			id = add_timer(hdl, ms, Topic::TEXT_INPUT, wrap(cmd_str));
-		}
-		else
-		{
-			id = delay(hdl, ms, Topic::TEXT_INPUT, wrap(cmd_str));
-		}
-		if (id == -1)
-		{
-			msa::agent::say(hdl, "Oh no! I'm sorry, $USER_TITLE, that didn't work quite right!");
-		}
-		else
-		{
-			msa::agent::say(hdl, "Okay, $USER_TITLE, I will do that " + type + " " + std::to_string(ms.count()) + " millisecond" + plural + "!");
-			msa::agent::say(hdl, "The timer ID is " + std::to_string(id) + ".");
-		}
-	}	
-
-	static void cmd_deltimer(msa::Handle hdl, const msa::cmd::ParamList &params, HandlerSync *const UNUSED(sync))
-	{
-		if (params.arg_count() < 1)
-		{
-			msa::agent::say(hdl, "Ahh... $USER_TITLE, I need to know which timer I should delete.");
-			return;
-		}
-		int16_t id = 0;
-		try
-		{
-			id = std::stoi(params[0]);
-		}
-		catch (std::exception &e)
-		{
-			msa::agent::say(hdl, "Sorry, $USER_TITLE, but '" + params[0] + "' isn't an integer.");
-			return;
-		}
-		remove_timer(hdl, id);
-		msa::agent::say(hdl, "Okay! I stopped timer " + std::to_string(id) + " for you, $USER_TITLE.");
 	}
 
 } }
