+++ src/event/hooks.hpp
@@ -12,9 +12,3 @@
 MSA_MODULE_HOOK(void, subscribe, msa::Handle msa, Topic topic, EventHandler handler)
 MSA_MODULE_HOOK(void, unsubscribe, msa::Handle msa, Topic topic, EventHandler handler)
 MSA_MODULE_HOOK(void, generate, msa::Handle msa, const Topic topic, const IArgs &args)
-MSA_MODULE_HOOK(int16_t, schedule, msa::Handle msa, time_t timestamp, const Topic topic, const IArgs &args)
-MSA_MODULE_HOOK(int16_t, delay, msa::Handle msa, std::chrono::milliseconds delay, const Topic topic, const IArgs &args)
-MSA_MODULE_HOOK(int16_t, add_timer, msa::Handle msa, std::chrono::milliseconds period, const Topic topic, const IArgs &args)
-MSA_MODULE_HOOK(void, remove_timer, msa::Handle msa, int16_t id)
-MSA_MODULE_HOOK(void, get_timers, msa::Handle msa, std::vector<int16_t> &list)
-
